refactor(efi): own console state in the console protocol module
`ConsoleState`, `InputState`, and `ScreenMode` move next to the text
input/output protocols that use them. Other protocol modules go through
`console::state()` and `console::with_state_mut()`.
